+++ src/components/auth/JoinComponent.js
@@ -1,8 +1,8 @@
-import { joinUser } from '../../business/auth';
-import router, { ROUTE_NAME } from '../../lib/router';
-import { showToast } from '../../lib/toast';
-import globalStore from '../../stores/globalStore';
-import { GLOBAL_STATE_KEYS } from '../../utils/constants';
+import { joinUser } from '../../business/logic/auth.js';
+import router, { ROUTE_NAME } from '../../lib/router.js';
+import { showToast } from '../../lib/toast.js';
+import globalStore from '../../business/store/globalStore.ts';
+import { GLOBAL_STATE_KEYS } from '../../utils/constants/index.ts';
 
 class JoinComponent {
   $app;
