+++ src/components/auth/LoginComponent.js
@@ -1,8 +1,8 @@
-import { loginUser } from '../../business/auth';
-import router, { ROUTE_NAME } from '../../lib/router';
-import { showToast } from '../../lib/toast';
-import globalStore from '../../stores/globalStore';
-import { GLOBAL_STATE_KEYS } from '../../utils/constants';
+import { loginUser } from '../../business/logic/auth.js';
+import router, { ROUTE_NAME } from '../../lib/router.js';
+import { showToast } from '../../lib/toast.js';
+import globalStore from '../../business/store/globalStore.ts';
+import { GLOBAL_STATE_KEYS } from '../../utils/constants/index.ts';
 
 class LoginComponent {
   $app;
