+++ src/components/vendingMachine/common/CoinTableComponent.js
@@ -1,7 +1,7 @@
-import { returnCoin } from '../../../business/vendingMachine';
-import { showToast } from '../../../lib/toast';
-import vendingMachineStore from '../../../stores/vendingMachineStore';
-import { VENDING_MACHINE_STATE_KEYS } from '../../../utils/constants';
+import { returnCoin } from '../../../business/logic/vendingMachine.js';
+import { showToast } from '../../../lib/toast.js';
+import vendingMachineStore from '../../../business/store/vendingMachineStore.ts';
+import { VENDING_MACHINE_STATE_KEYS } from '../../../utils/constants/index.ts';
 
 class CoinTableComponent {
   constructor($parent, { tableId, tableCaption }) {
