+++ src/utils/constants/error.ts
@@ -1,0 +1,14 @@
+export const ERROR_MSG = {
+  NAME_OVER_LIMIT_LENGTH: '상품 명은 10글자를 초과할 수 없습니다.',
+  NAME_EMPTY: '상품 명을 최소 한 글자를 입력해주세요.',
+  PRICE_OUT_OF_RANGE: '상품가격을 100이상 10000이하로 입력해주세요.',
+  PRICE_NOT_DIVISIBLE_BY_10: '상품가격을 10으로 나누어 떨어지도록 입력해주세요.',
+  QUANTITY_OUT_OF_RANGE: '상품 수량을 0이상 20이하의 수로 입력해주세요.',
+  PRODUCT_SOLD_OUT: '상품이 매진되었습니다.',
+  CAN_NOT_REFERENCE_STATE: '조회할 수 없는 상태값 입니다',
+  CHANGE_NOT_DIVISIBLE_BY_10: '잔돈은 10으로 나누어 떨어져야합니다.',
+  CHANGE_OUT_OF_RANGE: '잔돈은 10이상 100000이하의 금액을 투입하여야 합니다.',
+  NOT_ENOUGH_MONEY: '돈이 부족합니다.',
+  MACHINE_DONT_HAVE_MONEY: '자판기에 동전이 없습니다 ㅠㅠ 😱',
+  YOU_DONT_HAVE_MONEY: '사용자님 동전을 넣으셔야죠!',
+};
