+++ src/utils/constants/range.ts
@@ -1,0 +1,20 @@
+import { TRange } from '../types';
+
+export const NAME_LENGTH_LIMIT = 10;
+
+export const PRICE_RANGE: TRange = {
+  MIN: 100,
+  MAX: 10000,
+};
+
+export const MONEY_DIVIDE_STANDARD = 10;
+
+export const QUANTITY_RANGE: TRange = {
+  MIN: 0,
+  MAX: 20,
+};
+
+export const CHANGE_RANGE: TRange = {
+  MIN: 10,
+  MAX: 100000,
+};
