+++ src/utils/validation/vendingMachine.ts
@@ -1,13 +1,12 @@
+import { ERROR_MSG } from '../constants/error';
 import {
   CHANGE_RANGE,
-  ERROR_MSG,
   MONEY_DIVIDE_STANDARD,
   NAME_LENGTH_LIMIT,
   PRICE_RANGE,
   QUANTITY_RANGE,
-} from './constants';
+} from '../constants/range';
 
-/** Domain */
 export const isOverLimitLength = (nameInput: string) => nameInput.length > NAME_LENGTH_LIMIT;
 
 export const isEmptyName = (nameInput: string) => nameInput.length === 0;
@@ -56,29 +55,6 @@ export const checkChangeInput = (changeInput: number) => {
   }
   if (isNotdivisibleBy10(changeInput)) {
     throw new Error(ERROR_MSG.CHANGE_NOT_DIVISIBLE_BY_10);
-  }
-  return true;
-};
-
-/** User */
-export const isOutOfNameRange = name => {
-  return name.length < 2 || name.length > 6;
-};
-
-export const isInValidPassword = password => {
-  const regExp = /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/;
-  return password.match(regExp) === null;
-};
-
-export const checkJoinPossible = (name, password, passwordReenter) => {
-  if (isOutOfNameRange(name)) {
-    throw new Error('이름은 2자~6자 입니다.');
-  }
-  if (isInValidPassword(password)) {
-    throw new Error('비밀번호는 문자와 숫자 포함 8자 이상이어야 합니다.');
-  }
-  if (password !== passwordReenter) {
-    throw new Error('다시 입력한 비밀번호가 일치하지 않습니다.');
   }
   return true;
 };
